requirements.txt.txt
numpy
Pillow
moviepy
phidata
python-dotenv
SpeechRecognition
gTTS
requests
opencv-python